Tested CheckValidity and CustomException
