+++ src/main/java/com/springboot/parkinglot/controller/CheckValidity.java
@@ -1,5 +1,6 @@
 package com.springboot.parkinglot.controller;
 
+import com.springboot.parkinglot.common.CustomException;
 import com.springboot.parkinglot.controller.user.UserDto;
 
 public interface CheckValidity {
@@ -8,5 +9,5 @@ public interface CheckValidity {
 
     void checkLong(Long value);
 
-    void check(String name);
+    void check(String id, String password, String name) throws CustomException;
 }
